+++ JotunnLib/Entities/CustomLocation.cs
@@ -107,7 +107,7 @@ namespace Jotunn.Entities
         ///     Helper method to determine if a location prefab with a given name is a custom location created with Jötunn.
         /// </summary>
         /// <param name="prefabName">Name of the prefab to test.</param>
-        /// <returns>true if the prefab is added as a custom piece to the <see cref="PieceManager"/>.</returns>
+        /// <returns>true if the prefab is added as a custom location to the <see cref="ZoneManager"/>.</returns>
         public static bool IsCustomLocation(string prefabName)
         {
             return ZoneManager.Instance.Locations.ContainsKey(prefabName);